use create_app format to enable testing
